Add fixes to the database migration code
The script "fix-duplicate-device-uids.py" is
only needed once when we migrate the database
from the pre-update-12.sql schema to the new
one, as update-12.sql already includes the
necessary constraints to stop this bug from
happening.

Also make sure that running update-12.sql is
not a problem if the substitution table exists.