Conversion of EpisodeAction to CouchDB backend
The conversion is via the management command migrate-episode-actions and
should be carried out incrementally based on the Ids of the actions in the
relational backend (ids given with --min-id=<id> and --max-id=<id>).

As episode actions are never changed, the migration can catch up until it
reaches current items. If the backlog is small enough (eg a few thousand
objects), creation of new episode actions in the relational backend can
be stopped, all remaining objects migrated and the backend be switched to the
CouchDB based one (which still has to be implemented).