always return "released" attribute of episode_data
Return empty string for "released" attribute of episode_data when no value is set.  Follows the same logic as "podcast_url" among others.

I'm using the python client library and "released" is in REQUIRED_KEYS which generates an error when no key is present in response. I'm assuming that the preference would be this proposed fix, as opposed to fixing in the client.