Convert transparent logos to PNGs instead of JPEGs when resizing
RGBA images cannot be written as JPEGs. As a result, saving the resized
image fails, and the original image is sent as a fallback.

This is slow because it causes mygpo to try (and fail) to resize the
image on every request. It also breaks parts of the UI that expect a
smaller logo, not the large original.